# Reduce allocs LengthConverter to minimum and improve conversion performance