--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/DoubleCollection.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/DoubleCollection.cs
@@ -720,7 +720,7 @@ namespace System.Windows.Media
 
             while (th.NextToken())
             {
-                value = Convert.ToDouble(th.GetCurrentToken(), formatProvider);
+                value = double.Parse(th.GetCurrentTokenAsSpan(), formatProvider);
 
                 resource.Add(value);
             }