--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/DoubleCollection.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/DoubleCollection.cs
@@ -713,14 +713,14 @@ namespace System.Windows.Media
         {
             IFormatProvider formatProvider = System.Windows.Markup.TypeConverterHelper.InvariantEnglishUS;
 
-            TokenizerHelper th = new TokenizerHelper(source, formatProvider);
+            ValueTokenizerHelper tokenizer = new(source, formatProvider);
             DoubleCollection resource = new DoubleCollection();
 
             double value;
 
-            while (th.NextToken())
+            while (tokenizer.NextToken())
             {
-                value = double.Parse(th.GetCurrentTokenAsSpan(), formatProvider);
+                value = double.Parse(tokenizer.GetCurrentToken(), formatProvider);
 
                 resource.Add(value);
             }