--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/GradientStopCollection.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/GradientStopCollection.cs
@@ -805,7 +805,7 @@ namespace System.Windows.Media
             {
                 value = new GradientStop(
                     Parsers.ParseColor(th.GetCurrentToken(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider));
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider));
 
                 resource.Add(value);
             }