--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/PointCollection.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media/Generated/PointCollection.cs
@@ -721,8 +721,8 @@ namespace System.Windows.Media
             while (th.NextToken())
             {
                 value = new Point(
-                    Convert.ToDouble(th.GetCurrentToken(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider));
+                    double.Parse(th.GetCurrentTokenAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider));
 
                 resource.Add(value);
             }