--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Matrix3D.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Matrix3D.cs
@@ -215,7 +215,7 @@ namespace System.Windows.Media.Media3D
 
             Matrix3D value;
 
-            String firstToken = th.NextTokenRequired();
+            ReadOnlySpan<char> firstToken = th.NextTokenRequiredAsSpan();
 
             // The token will already have had whitespace trimmed so we can do a
             // simple string compare.
@@ -225,23 +225,22 @@ namespace System.Windows.Media.Media3D
             }
             else
             {
-                value = new Matrix3D(
-                    Convert.ToDouble(firstToken, formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider),
-                    Convert.ToDouble(th.NextTokenRequired(), formatProvider));
+                value = new Matrix3D(double.Parse(firstToken, formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
+                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider));
             }
 
             // There should be no more tokens in this string.