--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Matrix3D.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Matrix3D.cs
@@ -211,11 +211,11 @@ namespace System.Windows.Media.Media3D
         {
             IFormatProvider formatProvider = System.Windows.Markup.TypeConverterHelper.InvariantEnglishUS;
 
-            TokenizerHelper th = new TokenizerHelper(source, formatProvider);
+            ValueTokenizerHelper tokenizer = new(source, formatProvider);
 
             Matrix3D value;
 
-            ReadOnlySpan<char> firstToken = th.NextTokenRequiredAsSpan();
+            ReadOnlySpan<char> firstToken = tokenizer.NextTokenRequired();
 
             // The token will already have had whitespace trimmed so we can do a
             // simple string compare.
@@ -226,25 +226,25 @@ namespace System.Windows.Media.Media3D
             else
             {
                 value = new Matrix3D(double.Parse(firstToken, formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider),
-                    double.Parse(th.NextTokenRequiredAsSpan(), formatProvider));
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider),
+                    double.Parse(tokenizer.NextTokenRequired(), formatProvider));
             }
 
             // There should be no more tokens in this string.
-            th.LastTokenRequired();
+            tokenizer.LastTokenRequired();
 
             return value;
         }