--- a/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Point3D.cs
+++ b/src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Point3D.cs
@@ -157,7 +157,8 @@ namespace System.Windows.Media.Media3D
 
             ReadOnlySpan<char> firstToken = tokenizer.NextTokenRequired();
 
-            value = new Point3D(double.Parse(firstToken, formatProvider),
+            value = new Point3D(
+                double.Parse(firstToken, formatProvider),
                 double.Parse(tokenizer.NextTokenRequired(), formatProvider),
                 double.Parse(tokenizer.NextTokenRequired(), formatProvider));