… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Point3DCollection.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Point4D.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Quaternion.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Rect3D.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Size3D.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Vector3D.cs
… src/Microsoft.DotNet.Wpf/src/PresentationCore/System/Windows/Media3D/Generated/Vector3DCollection.cs
… src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/VirtualizationCacheLengthConverter.cs
… src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/CornerRadiusConverter.cs
… src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/LengthConverter.cs
… src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/ThicknessConverter.cs
… src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/TokenizerHelper.cs
… src/Microsoft.DotNet.Wpf/src/Shared/MS/Internal/ValueTokenizerHelper.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Generated/Int32Rect.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Generated/Point.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Generated/Rect.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Generated/Size.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Generated/Vector.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/System/Windows/Media/Generated/Matrix.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/ValueTokenizerHelper.cs
… src/Microsoft.DotNet.Wpf/src/WindowsBase/WindowsBase.csproj
… src/Microsoft.DotNet.Wpf/src/WpfGfx/codegen/mcg/generators/ManagedResource.cs
… src/Microsoft.DotNet.Wpf/src/WpfGfx/codegen/mcg/generators/ManagedStruct.cs
… src/Microsoft.DotNet.Wpf/src/WpfGfx/codegen/mcg/xml/Resource.xml