## Description

Simplifies code in `CanConvertTo`/`ConvertFrom`/`ConvertTo`, removes additional type casts by using `is` checks/pattern match. Commits should be easy to review.

Uses `ReadOnlySpan<char>` to go through the tokenized string input, improving performance whenever a unit is specified, the string needs to be trimmed, etc. I've opted for `OrdinalIgnoreCase` instead of lowercase (which should be an uppercase, really) conversion and then `Ordinal` compare as it removes an allocation.

The benchmarks use the stack-allocated version of `FromString` for the `PR_Chain` one (#9363). `Original` uses the original heap array alloc, hence the Allocated field could be adjusted with 256 bytes respectively on the `PR_Chain` results but it is still an improvement (there's no measurable time difference in that function, so I've opted for the complete call-chain testing to simplify testing). Multiple passes due to the for loop.

### Mixed use-cases with string manipulation (units)

```c#
CultureInfo CachedNorwegian = new CultureInfo("nn-NO");

FromString("10pt, 10pt", CultureInfo.InvariantCulture);
FromString("12px; 12px; 12px; 12px", CachedNorwegian);
FromString("10pt, 10pt", CultureInfo.InvariantCulture);
FromString("12px; 12px; 12px; 12px", CachedNorwegian);
FromString("10pt, 10pt", CultureInfo.InvariantCulture);
FromString("12px; 12px; 12px; 12px", CachedNorwegian);
FromString("10pt, 10pt", CultureInfo.InvariantCulture);
FromString("12px; 12px; 12px; 12px", CachedNorwegian);
```
| Method             | Mean [ns]  | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|------------------- |-----------:|-----------:|------------:|-------:|--------------:|
| PR_Edit_Chain| 1,635.5 ns |    9.78 ns |     8.67 ns | 0.0687 |        1152 B |
| Original_Chain      | 1,792.3 ns |   12.85 ns |    10.73 ns | 0.1411 |        2368 B |

### Initialization from simple values

```c#
CultureInfo CachedNorwegian = new CultureInfo("nn-NO");

FromString("1,1", CultureInfo.InvariantCulture);
FromString("1,1,1,1", CultureInfo.InvariantCulture);
FromString("1;1", CachedNorwegian);
FromString("1;1;1;1", CachedNorwegian);
FromString("1,1", CultureInfo.InvariantCulture);
FromString("1,1,1,1", CultureInfo.InvariantCulture);
FromString("1;1", CachedNorwegian);
FromString("1;1;1;1", CachedNorwegian);
```

| Method             | Mean [ns]  | Error [ns] | StdDev [ns] | Gen0   | Allocated [B] |
|------------------- |-----------:|-----------:|------------:|-------:|--------------:|
| PR_Edit_Chain| 1,266.7 ns |   16.35 ns |    15.30 ns | 0.0572 |         960 B |
| Original_Chain      | 1,421.3 ns |   10.32 ns |     9.65 ns | 0.0839 |        1408 B |

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, basic set of assert testing.

```c#
CultureInfo CachedNorwegian = new CultureInfo("nn-NO");

Assert.AreEqual(FromString("1,1,1,1", CultureInfo.InvariantCulture), FromString2("1,1,1,1", CultureInfo.InvariantCulture));
Assert.AreEqual(FromString("1PX,1PX,1PX,1PX", CultureInfo.InvariantCulture), FromString2("1PX,1PX,1PX,1PX", CultureInfo.InvariantCulture));

Assert.AreEqual(FromString("1,1", CultureInfo.InvariantCulture), FromString2("1,1", CultureInfo.InvariantCulture));
Assert.AreEqual(FromString("1", CultureInfo.InvariantCulture), FromString2("1", CultureInfo.InvariantCulture));

Assert.AreEqual(FromString("Auto", CultureInfo.InvariantCulture), FromString2("Auto", CultureInfo.InvariantCulture));
Assert.AreEqual(FromString("12px; 12px; 12px; 12px", CachedNorwegian), FromString2("12px; 12px; 12px; 12px", CachedNorwegian));

Assert.ThrowsException<FormatException>(() => FromString2("1,1,1", CultureInfo.InvariantCulture));
Assert.ThrowsException<FormatException>(() => FromString2("Auto, Auto, Auto", CultureInfo.InvariantCulture));
Assert.ThrowsException<FormatException>(() => FromString2(string.Empty, CultureInfo.InvariantCulture));

Assert.ThrowsException<FormatException>(() => FromString2("1,1,1", CachedNorwegian));
Assert.ThrowsException<FormatException>(() => FromString2("Auto, Auto, Auto", CachedNorwegian));
Assert.ThrowsException<FormatException>(() => FromString2(string.Empty, CachedNorwegian));
```

## Risk

Low.