@h3xds1nz 2025-03-24
@himgoyalmicro I've noticed this has been cut, can you share details about the failed test(s)?

This is really important PR and preferrably it should be in .NET 10.